+++ be/main.py
@@ -1,11 +1,9 @@
-import json
-from schedule.service import ScheduleService
+from contextlib import asynccontextmanager
 
 from fastapi.middleware.cors import CORSMiddleware
-
-from contextlib import asynccontextmanager
 from fastapi import FastAPI, HTTPException
-import os
+
+from game_schedule.service import ScheduleService
 
 
 schedule_service = ScheduleService()
